requirements.txt
gunicorn
dash
dash-bootstrap-components
dash-core-components
pandas
plotly-express
dash-daq
dash-html-components
numpy
xlrd
dash-extensions
pybase64
